Do challenge 18/Part 2

+++ AdventOfCode.Console/Models/LavaductLagoon.cs
@@ -1,6 +1,6 @@
 namespace AdventOfCode.Console.Models
 {
-    public record DigCommand(CardinalDirection Direction, int NumSteps);
+    public record DigCommand(CardinalDirection Direction, long NumSteps);
     public class LavaductLagoon
     {
         private readonly IEnumerable<DigCommand> _digPlan;
@@ -10,7 +10,7 @@ namespace AdventOfCode.Console.Models
             _digPlan = digPlan;
         }
 
-        private static (int, int) NewCorner((int, int) previousCorner, CardinalDirection direction, int numSteps)
+        private static (long, long) NewCorner((long, long) previousCorner, CardinalDirection direction, long numSteps)
         {
             return direction switch
             {
@@ -21,29 +21,30 @@ namespace AdventOfCode.Console.Models
             };
         }
 
-        public int Volume()
+        public long Volume()
         {
-            List<(int, int)> corners = new() { (0, 0) };
-            int numPointsOnEdge = 1;
+            List<(long, long)> corners = new() { (0, 0) };
+            long numPointsOnEdge = 1;
             foreach (DigCommand digCommand in _digPlan)
             {
-                (int, int) lastCorner = corners[^1];
-                (int, int) newCorner = NewCorner(lastCorner, digCommand.Direction, digCommand.NumSteps);
+                (long, long) lastCorner = corners[^1];
+                (long, long) newCorner = NewCorner(lastCorner, digCommand.Direction, digCommand.NumSteps);
                 corners.Add(newCorner);
                 numPointsOnEdge += digCommand.NumSteps;
             }
 
             // Shoelace formula
-            int twiceArea = 0;
+            long twiceArea = 0;
             for (int i = 0; i < corners.Count - 1; i++)
             {
-                (int x1, int y1) = corners[i];
-                (int x2, int y2) = corners[i + 1];
+                (long x1, long y1) = corners[i];
+                (long x2, long y2) = corners[i + 1];
                 twiceArea += x1 * y2 - x2 * y1;
             }
 
             double twicePoints = Math.Abs(twiceArea) + 1 + numPointsOnEdge;
-            return (int)Math.Ceiling(twicePoints / 2);
+
+            return (long)Math.Ceiling(twicePoints / 2);
         }
     }
 }
